:sparkles: Implement CloudMusic API & VLC player

+++ entities/__init__.py
@@ -1,7 +1,8 @@
 import os
-from interface.led import LED as _LED
 from interface.azure import parse_wav_file, speech_to_text, text_to_speech
-from interface.audio import play_wav_file
+from interface.audio import VLCInterface
+from interface.cloudmusic import CloudMusic
+from interface.led import LED as _LED
 from interface.qingyunke import get_reply
 from utils.log import init_logging
 
@@ -13,9 +14,16 @@ class Sory:
         self.detector = detector
         self.config = config
         self.LED = led
+        self.VLC_instance = VLCInterface()
+        self.music_provider = CloudMusic(self.config.netease_api_url)
+        self.music_provider.login_by_email(
+            self.config.netease_email, md5_password=self.config.netease_password_md5)
+        self.is_playing = False
 
-    def detected_callback(self, fname: str = ""):
-        play_wav_file("audio/wozai.wav")
+    def detected_callback(self):
+        if self.is_playing:
+            return
+        self.VLC_instance.play_audio("audio/wozai.wav")
         self.LED.power.on()
         for i in range(0, 12):
             self.LED.switch_by_place(
@@ -23,6 +31,8 @@ class Sory:
         logger.info('Recording audio...')
 
     def audio_recorder_callback(self, fname):
+        if self.is_playing:
+            return
         logger.info("Converting audio to text")
         try:
             audio_data = parse_wav_file(fname)
@@ -34,19 +44,32 @@ class Sory:
                 "Could not request results from STT service; {0}".format(e))
             reply = "语音转文字功能出错啦！"
         else:
-            reply = get_reply(stt_text)
-            logger.info("Bot: " + reply)
+            parsed_str = stt_text.split("点歌")
 
-        try:
-            play_back_data = text_to_speech(
-                reply, self.config.azure_key, "japaneast", "playback_" + fname)
-        except Exception as e:
-            logger.warning(
-                "Could not request results from TTS service; {0}".format(e))
-        else:
-            play_wav_file("playback_" + fname)
+            self.is_playing = False
+            if len(parsed_str) > 1:
+                logger.info("Trying to play: " + parsed_str[1])
+                song_list = self.music_provider.search_music(parsed_str[1])
+                if song_list != None and len(song_list) > 0:
+                    song_url = self.music_provider.get_song_url(song_list[0])
+                    self.VLC_instance.play_audio(song_url)
+                    self.is_playing = True
+
+            if self.is_playing == False:
+                reply = get_reply(stt_text)
+                logger.info("Bot: " + reply)
+                try:
+                    play_back_data = text_to_speech(
+                        reply, self.config.azure_key, "japaneast", "playback_" + fname)
+                except Exception as e:
+                    logger.warning(
+                        "Could not request results from TTS service; {0}".format(e))
+                else:
+                    self.VLC_instance.play_audio("playback_" + fname)
+                    self.is_playing = True
+                finally:
+                    os.remove("playback_" + fname)
         finally:
             os.remove(fname)
-            os.remove("playback_" + fname)
 
         self.LED.power.off()
